+++ pub-crawler/bin/add_seeds.py
@@ -29,8 +29,7 @@ async def add_seeds(input_filename, r, *, transport=None):
                 if not wf:
                     continue
                 job = {"job_type": "webfinger", "webfinger": wf}
-                if not await dispatcher.seen(job):
-                    await dispatcher.enqueue(job)
+                await dispatcher.enqueue_if_unseen(job)
 
     finally:
         await wfc.aclose()